Ticket #1 Initial Commit of PDF Scraping Utilities
We currently do not have established standards for commit messages for this project, so please forgive any blunders.

This patch works on Ticket #1 (see https://github.com/dance-score-sheets/dance-score-sheets/issues/1).

This patch includes some basic utilities for converting PDFs. They work without problems on the ILHC score sheets I've tried them with so far (but we don't have "real" tests yet, see the ticket for more info wrt that).

It is in Python2 rather than Python3, which we as a team decided on, so I'll address that in future commits.

/dance-score-sheets/misc_utilities/pdf_scraping_utilities.py
A CONVERT_ALL_PDFS_TO_CSVS
A PDF_TO_CSV
A FILE_HAS_EXTENSION
A DWIM_FILE_EXTENSION